requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==5.1.5
django-extensions==3.2.3
feedparser==6.0.11
pillow==11.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
